+++ pkg/db/driver_oracle.go
@@ -1,0 +1,7 @@
+//go:build oracle
+
+// Package db: Oracle driver registration.
+// Activate with: go build -tags oracle
+package db
+
+import _ "github.com/sijms/go-ora/v2"